requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
defusedxml==0.7.1
dict2xml==1.7.1
dicttoxml==1.7.4
Django==3.2.12
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
idna==3.2
importlib-metadata==4.8.3
json2xml==3.16.0
Markdown==3.3.6
python-dotenv==0.20.0
pytz==2022.1
requests==2.26.0
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.7
xmltodict==0.12.0
zipp==3.6.0
